Criando arquivo models para refatorar o codigo
